UnfoldMixedModels.UnfoldLinearMixedModel — Type

Concrete type to implement an Mass-Univariate LinearMixedModel. .design contains the formula + times dict .designmatrix contains a DesignMatrixmodelfit is a Any container for the model results

UnfoldMixedModels.UnfoldLinearMixedModelContinuousTime — Type

Concrete type to implement an deconvolution LinearMixedModel.

Warning This is to be treated with care, not much testing went into it.

.design contains the formula + times dict .designmatrix contains a DesignMatrix.modelfit is a Any container for the model results